+++ CHANGELOG.md
@@ -3,3 +3,7 @@
 ## [Unreleased]
 
 -   Initial release
+
+## [2025-10-03]
+
+-   Changed from using nx cli to list the projects to using the vscode api to find the projects ourself